component NavBarMenu.tsx added navigation added
