+++ package.json
@@ -26,7 +26,6 @@
     "@uiw/react-color": "^2.3.0",
     "axios": "^1.7.2",
     "date-fns": "^3.6.0",
-    "globals": "^15.6.0",
     "js-cookie": "^3.0.5",
     "jwt-decode": "^4.0.0",
     "react": "^18.2.0",
@@ -40,7 +39,6 @@
     "zustand": "^4.5.2"
   },
   "devDependencies": {
-    "@pnpm/eslint-config": "^1.1.0",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
@@ -57,7 +55,6 @@
     "eslint-plugin-unused-imports": "^4.0.0",
     "prettier": "^3.2.5",
     "prettier-eslint": "^16.3.0",
-    "tslint": "^6.1.3",
     "typescript": "^5.2.2",
     "vite": "^5.2.0"
   },
